chore: 清理根目录，建立项目结构规范
- 移动构建日志到 temp/build_logs/
- 移动错误分析文件到 temp/analysis/
- 移动报告文档到 docs/reports/
- 移动计划文档到 docs/plans/
- 更新 .gitignore 排除临时文件
- 创建文档导航 docs/README.md

+++ kernel/src/sync/futex_tests.rs
@@ -335,8 +335,8 @@ impl FutexTestSuite {
         let avg_latency = total_time / operations;
         
         // Update stats
-        self.stats.total_operations += operations;
-        self.stats.successful_operations += operations;
+        self.stats.total_operations += operations as usize;
+        self.stats.successful_operations += operations as usize;
         self.stats.avg_latency_ns = avg_latency;
         
         crate::println!("[futex_test] Performance test passed, avg latency: {}ns", avg_latency);
